Change git branch when patches are applied.
Apply patches on a separate branch. The separate branch is
automatically created when Guilt pushes something, and removed when no
patches are applied. The name is formed by prepending "guilt/" to the
original branch.

This breaks the "upstream" relationship, so a mistaken "git push"
while patches are applied will no longer mess up your upstream
repository.

Update the testsuite and documentation.

Thanks to Junio C Hamano for suggesting this solution to my problem.

Signed-off-by: Per Cederqvist <cederp@opera.com>