disp no longer processes backslashes.
Only one invocation of "disp" or "_disp" actually needed backslash
processing.  In quite a few instances, it was wrong to do backslash
processing.  Created the new function "disp_e" that should be used
when backslash processing is required, and changed "disp" and "disp_"
to use printf code %s instead of "%b".