disp no longer processes backslashes.
Only one invocation of "disp" or "_disp" actually needed backslash
processing.  In quite a few instances, it was wrong to do backslash
processing, as the message contained data derived from the user.

Created the new function "disp_e" that should be used when backslash
processing is required, and changed "disp" and "disp_" to use printf
code %s instead of "%b".

Signed-off-by: Per Cederqvist <cederp@opera.com>